Use the correct scope for options
Set all the options in their correct scope using set_option

Options need to be set in the correct scope
(e.g. buffer, window or global) in order to work.
`:set` in vim handles this automatically but in lua this
needs to be done manually.

To help with this set_options sets the option in the given scope and
global scope as well.

To find out which scope an option should be set in see `:help <option>`