Create keyboard shortcut to toggle compe (completion)
Sometimes the completion get's in the way (e.g. in markdown files -
where it is turned of by default) in that case I quickly want to turn
off the completion.

In markdown files completion is turned off by default but sometimes I
would like the completion. So it would help if I could toggle it on/off.