Fix statusline to show correct filenames
The filename of the file in the active buffer was shown in the
statusline of the active buffer but also of the not active buffers.
This happened because in `get_current_filename` the current active
filename was returned with the icon. But this function is also used for
the inactive filename. So instead of using the filename from the current
buffer the built in statusline `%t` can be used to show the filename of
each window/panel.