Create a function to toggle the nvim file tree
nvim-tree.lua does not provide a toggle function to use with
`:NvimTreeFindFile`. So opening and closing the tree cannot be bound to
the same key when using the find file functionality. With the custom
function the toggling behaviour can be accomplished with one keybind and
the find file functionality is kept.